fix 2-4
what a stupid for me, why i use the loop

+++ 计算机程序设计（C++）西安交通大学/Week2/2-3数据的简单统计.cpp
@@ -7,7 +7,7 @@ int main(int argc, const char * argv[]) {
 	long sum,avg;
 	double avg_d;
 	sum=a+b+c;
-	avg_d=sum/3;
+	avg_d=(double)sum/3;
 	avg=int(avg_d+0.5);
 	cout<<sum<<endl;
 	cout<<avg_d<<endl;
